@@ -128,7 +128,7 @@ describe('ClientRedis', () => {
 
         subscription = client.createResponseCallback();
         client['routingMap'].set(responseMessage.id, callback);
-        subscription('channel', new Buffer(JSON.stringify(responseMessage)));
+        subscription('channel', Buffer.from(JSON.stringify(responseMessage)));
       });
       it('should call callback with expected arguments', () => {
         expect(
@@ -146,7 +146,7 @@ describe('ClientRedis', () => {
         client['routingMap'].set(responseMessage.id, callback);
         subscription(
           'channel',
-          new Buffer(
+          Buffer.from(
             JSON.stringify({
               ...responseMessage,
               isDisposed: responseMessage.response,
@@ -170,7 +170,7 @@ describe('ClientRedis', () => {
       beforeEach(() => {
         callback = sinon.spy();
         subscription = client.createResponseCallback();
-        subscription('channel', new Buffer(JSON.stringify(responseMessage)));
+        subscription('channel', Buffer.from(JSON.stringify(responseMessage)));
       });
 
       it('should not call callback', () => {
